// src/app/models/categoria.ts
export interface Random {
  p?: any;
  p1?: string;
  p2?: any;
  p3?: any;
  p4?: any;
  p5?: any;
  p6?: any;
  p7?: any;
  p8?: any;
  p9?: any;
  p10?: any;
  p11?: any;
  p12?: any;

  par1?: any;
  par2?: any;

  np?: any;
  np1?: string;
  np2?: any;
  np3?: any;
  np4?: any;
  np5?: any;
  np6?: any;
  np7?: any;
  np8?: any;
  np9?: any;
  np11?: any;
  np12?: any;
  np13?: any;

  descripcion?: string;
  descripcion1?: string;
  descripcion2?: string;
  descripcion3?: string;
  descripcion4?: string;
  descripcion5?: string;
  descripcion6?: string;
  descripcion7?: string;

  img?: string;
  img1?: string;
  img2?: string;
  img3?: string;
  img4?: string;
  img5?: string;
  img6?: string;
  img7?: string;
  img8?: string;
  img9?: string;


  id?: string;




  createdDate: any;
  lastModifiedDate: any;

  titlePost: string;
  contentPost: string;
  imagePost?: any;
  tagsPost: string;
  fileRef?: string;


}

export interface Memes {
  p?: any;
  p1?: string;
  p2?: any;
  p3?: any;
  p4?: any;
  p5?: any;
  p6?: any;
  p7?: any;
  p8?: any;
  p9?: any;
  p10?: any;
  p11?: any;
  p12?: any;

  par1?: any;
  par2?: any;

  np?: any;
  np1?: string;
  np2?: any;
  np3?: any;
  np4?: any;
  np5?: any;
  np6?: any;
  np7?: any;
  np8?: any;
  np9?: any;
  np11?: any;
  np12?: any;
  np13?: any;

  descripcion?: string;
  descripcion1?: string;
  descripcion2?: string;
  descripcion3?: string;
  descripcion4?: string;
  descripcion5?: string;
  descripcion6?: string;
  descripcion7?: string;

  img?: string;
  img1?: string;
  img2?: string;
  img3?: string;
  img4?: string;
  img5?: string;
  img6?: string;
  img7?: string;
  img8?: string;
  img9?: string;


  id?: string;




  createdDate: any;
  lastModifiedDate: any;

  titlePost: string;
  contentPost: string;
  imagePost?: any;
  tagsPost: string;
  fileRef?: string;


}

export interface FileI {
  name: string;
  imageFile: File;
  size: string;
  type: string;
  createdDate: any;
  lastModifiedDate: any;

}